fix(app-ui): CodeRabbit 리뷰 반영 — 당류 double·총계 폴백·에셋/차트 안전성
- diet: DietEntry/DietDay 당류를 double로 통일해 8.5·14.8 소수 유지(수정 시트·repository 포함)
- diet: per-food 합산이 0이면 entry/day 총계로 폴백(실서버 응답에서 0 방지)
- diet: 식사 썸네일 Image.asset errorBuilder로 이모지 폴백
- dashboard: 홈 당류 말풍선 소수 1자리 유지(14.8) + 그림자 주석 실제 값과 일치
- exercise: 도넛 0분 세그먼트 가드, 막대 폭 슬롯 비례 상한, 오늘 도넛 FittedBox로 오버플로 방지

+++ frontend/flutter/lib/features/diet/domain/entities/diet_day.dart
@@ -45,7 +45,7 @@ class DietEntry {
   final List<FoodItem> foods;
   final int totalCalories;
   final int sodiumMg;
-  final int sugarG;
+  final double sugarG;
 
   /// Short per-meal AI feedback shown on the diet-tab meal card. Empty when the
   /// backend hasn't produced a per-entry comment.
@@ -65,7 +65,7 @@ class DietEntry {
         .toList(),
     totalCalories: (json['total_calories']! as num).toInt(),
     sodiumMg: (json['sodium_mg'] as num?)?.toInt() ?? 0,
-    sugarG: (json['sugar_g'] as num?)?.toInt() ?? 0,
+    sugarG: (json['sugar_g'] as num?)?.toDouble() ?? 0,
     aiComment: (json['ai_comment'] as String?) ?? '',
   );
 }
@@ -102,7 +102,7 @@ class DietDay {
   final int totalCalories;
   final DietMacros macros;
   final int totalSodiumMg;
-  final int totalSugarG;
+  final double totalSugarG;
   final String aiCoachMessage;
 
   factory DietDay.fromJson(Map<String, Object?> json) => DietDay(
@@ -112,7 +112,7 @@ class DietDay {
         .toList(),
     totalCalories: (json['total_calories']! as num).toInt(),
     totalSodiumMg: (json['total_sodium_mg']! as num).toInt(),
-    totalSugarG: (json['total_sugar_g']! as num).toInt(),
+    totalSugarG: (json['total_sugar_g']! as num).toDouble(),
     macros: DietMacros.fromJson(json['macros']! as Map<String, Object?>),
     aiCoachMessage: json['ai_coach_message']! as String,
   );
